refactor template hooks
